requirements.txt
google-generativeai
edge-tts
pygame
SpeechRecognition
pyaudio
customtkinter
pywhatkit
pyautogui
psutil
keyboard
playsound
wikipedia
